Fixed --add functionality, improved startup checks
--add functionality from the command line now works with better parsing.
It's word-based rather than character-based.

The syntax for both the quick-add (from CLI) and add from VIM is kinda
rough. Maybe trying to build as a string rather than individual lines in
an array would be easier?

Definitely would be when reading back. Wouldn't have to iterate
backwards so many times.

TODO:
   - Fix issue in which newline is prepended to the --read output